Add separate os_deps container and artifact/ dir
The os_deps container is a separate container for preparing "OS dependencies".
These are things that Go binaries require to run correctly. Specifically, it
prepares a zipped-up db of timezone data, used by the time.LoadLocation() func,
and an up-to-date collection of CA certs, required for HTTPS connections.

Sometimes we would like to access the Go binary that's built inside the build
container. Running ./artifact/build.sh will cause the build container to output
the Go binary to artifact/